refactor: DF-380 - Move enums to separate types file
