+++ src/server/plugins/engine/types/enums.ts
@@ -1,0 +1,15 @@
+export enum UploadStatus {
+  initiated = 'initiated',
+  pending = 'pending',
+  ready = 'ready'
+}
+
+export enum FileStatus {
+  complete = 'complete',
+  rejected = 'rejected',
+  pending = 'pending'
+}
+
+export enum FormAdapterSubmissionSchemaVersion {
+  V1 = 1
+}
